Refactor exact predicates class into functions.
